+++ mujoco_ros/CHANGELOG.md
@@ -1,3 +1,9 @@
+<a name="Unreleased"></a>
+## [Unreleased]
+
+### Fixed
+* re-added setting realtime settings via ros param or in the mujoco model xml.
+
 <a name="0.7.0"></a>
 ## [0.7.0] - 2023-08-15
 
@@ -204,6 +210,7 @@ Contributors: @DavidPL1, @balandbal
 
 Contributors: @DavidPL1
 
+[Unreleased]: https://github.com/ubi-agni/mujoco_ros_pkgs/compare/0.7.0...HEAD
 [0.7.0]: https://github.com/ubi-agni/mujoco_ros_pkgs/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/ubi-agni/mujoco_ros_pkgs/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/ubi-agni/mujoco_ros_pkgs/compare/0.4.0...0.5.0